+++ ledger/api/managers/template_manager.py
@@ -355,6 +355,9 @@ class TemplateProcess:
                         "total_amount_day": (
                             round(amounts[key]["total_amount_day"], 2)
                         ),
+                        "total_amount_day_tick": (
+                            round(amounts[key]["total_amount_day"] / 3, 2)
+                        ),
                         "total_amount_hour": (
                             round(amounts[key]["total_amount_hour"], 2)
                             if key != "mining"
@@ -366,6 +369,9 @@ class TemplateProcess:
                         ),
                         "average_hour": round(
                             (amounts[key]["total_amount"] / current_day) / 24, 2
+                        ),
+                        "average_hour_tick": round(
+                            (amounts[key]["total_amount"] / current_day) / 24 / 3, 2
                         ),
                     }.items()
                     if value != 0