+++ ledger/tests/test_api/test_char_template_api.py
@@ -96,4 +96,4 @@ class ManageApiTemplateCharEndpointsTest(TestCase):
 
         response = self.client.get(url)
 
-        self.assertEqual(response.status_code, 403)
+        self.assertContains(response, "Permission Denied", status_code=200)